Features deleted between polling are not captured as deletes
If you are using polling and a feature has been deleted (or retired)
between polls (or between SSE refreshes), then the SDK would not pick up
the delete happened and "exists()" would return true. This is also true for
catch and release, if a feature because non-existent between polling.

This change ensures that if a feature becomes "missing" in a full feature
list update from Edge, it will ensure they get marked as deleted.